pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Егор <oegor8194@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.6"
gunicorn = "^20.0.4"
python-dotenv = "^0.15.0"
flake8 = "^3.8.4"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
django-bootstrap4 = "^2.3.1"
whitenoise = "^5.2.0"
psycopg2 = "^2.8.6"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
